MIM-202 统一工作区与会话列表行布局
